move compiled modules into src
