+++ mcdc/iqmc/iqmc_loop.py
@@ -4,9 +4,9 @@ from numpy import ascontiguousarray as cga
 from numba import njit, objmode
 
 import mcdc.adapt as adapt
-import mcdc.geometry as geometry
 import mcdc.iqmc.iqmc_kernel as iqmc_kernel
 import mcdc.kernel as kernel
+import mcdc.src.geometry as geometry
 import mcdc.local as local
 
 from mcdc.constant import *
