--- mcdc/mesh/__init__.py
@@ -1,1 +1,0 @@
-from mcdc.mesh import structured, uniform
