+++ mcdc/src/geometry.py
@@ -3,8 +3,8 @@ import math
 from numba import njit, int64
 
 import mcdc.local as local
-import mcdc.mesh as mesh
-import mcdc.physics as physics
+import mcdc.src.mesh as mesh
+import mcdc.src.physics as physics
 import mcdc.src.surface as surface_
 
 from mcdc.constant import *
