+++ mcdc/src/mesh/__init__.py
@@ -1,0 +1,1 @@
+from mcdc.src.mesh import structured, uniform
